+++ Assets/Game/Resources/Json/UseItemJson.json
@@ -1,8 +1,9 @@
 {
-  "itemname": [ "포션","비약","해독제" ],
-  "contents": [ "체력을 50채워주는 일반적인 포션","체력을 모두 채워주는 비약","독에걸린 상태를 회복해주는 약" ],
-  "value": [ 100,500,100 ],
-  "imageNum": [ 1,2,3 ],
-  "effectValue": [ 50,200,0 ],
-  "maxStack": [ 10,2,10 ]
+  "itemname": [ "포션", "비약", "해독제" ],
+  "contents": [ "체력을 50채워주는 일반적인 포션", "체력을 모두 채워주는 비약", "독에걸린 상태를 회복해주는 약" ],
+  "value": [ 100, 500, 100 ],
+  "imageNum": [ 1, 2, 3 ],
+  "effectValue": [ 50, 200, 0 ],
+  "maxStack": [ 10, 2, 10 ],
+  "useItemType": [ 0,0,0]
 }
