+++ Assets/Game/Scripts/Character/PlayerStatus.cs
@@ -5,6 +5,7 @@ using UnityEngine;
 public class PlayerStatus : MonoBehaviour
 {
     private int maxHp=100;
+    [SerializeField]
     private int hp;
     public int atk = 50;
     public int def = 0;
@@ -12,7 +13,13 @@ public class PlayerStatus : MonoBehaviour
     public int Hp
     {
         get { return hp; }
-        set { hp = value; }
+        set 
+        {
+            if (value > maxHp)
+                hp = maxHp;
+            else
+                hp = value; 
+        }
     }
     private void Awake()
     {
