+++ Assets/Game/Scripts/Item/BattleItemSystem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 526e6c989a2c02149bcf2e75386d2f49
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
