+++ Assets/Game/Scripts/Manager/InventoryManager.cs
@@ -15,7 +15,7 @@ public class InventoryManager : Singleton<InventoryManager>
     public Inventory inventory;
     public EquipmentInventory equipmentInventory;
     public PlayerStatus status;
-    public UseItemUI itemUI;
+    public BattleItemSystem itemUI;
 
     public int ItemCount
     {
@@ -49,13 +49,14 @@ public class InventoryManager : Singleton<InventoryManager>
     public bool AddItem(UseItem target)
     {
         bool addCount = false;
+        bool pass = false;
         for(int i = 0; i < itemcount; i++)
         {
             if(useItemList[i].itemName==target.itemName)
             {
                 if(useItemList[i].stack== useItemList[i].maxStack)
                 {
-                    addCount = true;
+                    pass = true;
                     break;
                 }
                 else
@@ -66,7 +67,7 @@ public class InventoryManager : Singleton<InventoryManager>
                 }
             }
         }
-        if(addCount == false)
+        if(addCount == false && pass == false)
         {
             for(int i = 0; i < useItemList.Length; i++)
             {
@@ -103,10 +104,10 @@ public class InventoryManager : Singleton<InventoryManager>
             }
             ItemCount--;
         }
-        if(inventory.gameObject.activeInHierarchy)
-        {
+        if (inventory.gameObject.activeInHierarchy)
             inventory.Refresh();
-        }
+        if (itemUI.gameObject.activeInHierarchy)
+            itemUI.SelectNum = itemUI.SelectNum;
     }
 
     public void SortItem(bool ascend = true)
@@ -138,6 +139,10 @@ public class InventoryManager : Singleton<InventoryManager>
         }
         for (int i = 0; i < ItemCount; i++)
             useItemList[i] = (tempList[i]);
+        if (inventory.gameObject.activeInHierarchy)
+            inventory.Refresh();
+        if (itemUI.gameObject.activeInHierarchy)
+            itemUI.SelectNum = itemUI.SelectNum;
     }
 
     public void Equip(EquipmentItem target)
